Rename components; improve FE structure
