+++ frontend/src/pages/alunos/AddNewAluno.js
@@ -1,7 +1,6 @@
 import React, { useState, useEffect } from 'react'
 import { Link, useNavigate } from 'react-router-dom'
 import InputMask from 'react-input-mask';
-import { TurmasData } from '../5Turmas/TurmasData';
 import { getAlunos, getTurmas, inserirDadosAlunos } from "../../service"
 
 const AddNewAluno = () => {
