+++ frontend/src/pages/turmas/TurmasData.js
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react'
+import { useEffect, useState } from 'react'
 import { getTurmas } from "../../service"
 
 export const TurmasData = () => {
