+++ frontend/src/routes/routes.js
@@ -1,20 +1,12 @@
 import React from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 
-
 const DefinedRoutes = () => {
     return (
         <BrowserRouter>
-
             <Routes>
 
-
-
-
-
-
             </Routes>
-
         </BrowserRouter>
     )
 }
